+++ Servers/UI/OJS.Servers.Ui/ClientApp/package.json
@@ -54,8 +54,8 @@
     "react-modal": "^3.14.3",
     "react-notifications-component": "^4.0.1",
     "react-redux": "^8.1.3",
-    "react-router": "^6.3.0",
-    "react-router-dom": "^6.3.0",
+    "react-router": "^6.24.0",
+    "react-router-dom": "^6.24.0",
     "react-scripts": "^5.0.1",
     "react-slick": "^0.29.0",
     "react-youtube": "^10.0.0",
