+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/portals/client/ClientPortal.module.scss
@@ -1,0 +1,9 @@
+@use 'src/styles/spacings';
+
+.main {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  padding-bottom: spacings.$sp-96;
+  padding-top: spacings.$sp-96;
+}
