--- Servers/UI/OJS.Servers.Ui/ClientApp/src/layout/content/PageContent.module.scss
@@ -1,9 +1,0 @@
-@use 'src/styles/spacings';
-
-.main {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  padding-bottom: spacings.$sp-96;
-  padding-top: spacings.$sp-96;
-}
