--- Servers/UI/OJS.Servers.Ui/ClientApp/src/layout/content/PageContent.tsx
@@ -1,11 +1,0 @@
-import { Route, Routes } from 'react-router-dom';
-
-import ClientPortal from '../../components/portals/client/ClientPortal';
-
-const PageContent = () => (
-    <Routes>
-        <Route path="/*" element={<ClientPortal />} />
-    </Routes>
-);
-
-export default PageContent;
