+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/pages/home/HomePage.tsx
@@ -11,6 +11,7 @@ import useTheme from '../../hooks/use-theme';
 import { useGetHomeStatisticsQuery } from '../../redux/services/homeStatisticsService';
 import { flexCenterObjectStyles } from '../../utils/object-utils';
 import { setLayout } from '../shared/set-layout';
+import withTitle from '../shared/with-title';
 
 import 'react-lite-youtube-embed/dist/LiteYouTubeEmbed.css';
 import styles from './HomePage.module.scss';
@@ -120,4 +121,4 @@ const HomePage = () => {
     );
 };
 
-export default setLayout(HomePage);
+export default setLayout(withTitle(HomePage, 'Home'));
